# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: DaYu Platform HTTP API
  version: 1.0.0-draft
  description: |
    大禹平台第一版 HTTP 契约。所有业务时间均为 UTC ISO-8601。
    浏览器先通过 GET /api/v1/session 获取与当前 Session 绑定的 CSRF Token；
    所有 POST、PUT、PATCH、DELETE 请求都必须携带 X-CSRF-TOKEN。
    本契约不暴露服务器绝对路径、NC 相对路径或 Nginx 内部下载位置。
servers:
  - url: /
...
paths:
  /api/v1/products:
    get:
      tags: [Catalog]
      summary: 查询已发布产品
      security: []
      responses:
        '200':
          description: 按 sortOrder 排序的产品列表，只包含已启用模式
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicProductSummary'
        '500': {$ref: '#/components/responses/InternalError'}
...
  /api/v1/preview-frames:
    get:
      tags: [Discovery]
      summary: 查询近期 WebP 预览帧
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductCodeQuery'
        - $ref: '#/components/parameters/DataModeQuery'
        - $ref: '#/components/parameters/FromRequired'
        - $ref: '#/components/parameters/ToRequired'
        - $ref: '#/components/parameters/CycleTimeOptional'
        - $ref: '#/components/parameters/LeadMinutesOptional'
      responses:
        '200':
          description: 按 validTime 正序排列；没有数据时 items 为空
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/PreviewFrame'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/forecast-cycles:
    get:
      tags: [Discovery]
      summary: 查询可用预报批次
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductCodeQuery'
        - name: assetType
          in: query
          required: true
          schema: {$ref: '#/components/schemas/AssetType'}
        - $ref: '#/components/parameters/FromRequired'
        - $ref: '#/components/parameters/ToRequired'
      responses:
        '200':
          description: 按 cycleTime 倒序排列
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/ForecastCycle'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/scientific-assets:
    get:
      tags: [Discovery]
      summary: 分页检索 NetCDF 科学数据
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductCodeQuery'
        - $ref: '#/components/parameters/DataModeQuery'
        - $ref: '#/components/parameters/FromRequired'
        - $ref: '#/components/parameters/ToRequired'
        - $ref: '#/components/parameters/CycleTimeOptional'
        - $ref: '#/components/parameters/LeadMinutesOptional'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 按 validTime 倒序排列的可用 NC 元数据
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScientificAssetPage'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/v1/admin/products:
    get:
      tags: [Catalog, Operations]
      summary: 管理员查询全部产品
      security: [{sessionCookie: []}]
      parameters:
        - {name: family, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/ProductStatus'}}
        - {name: code, in: query, schema: {$ref: '#/components/schemas/ProductCode'}}
      responses:
        '200':
          description: 管理产品列表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/AdminProduct'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Catalog, Operations]
      summary: 创建草稿产品
      security: [{sessionCookie: [], csrfToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateProductRequest'}
      responses:
        '201':
          description: 草稿产品
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/products/{productId}:
    put:
      tags: [Catalog, Operations]
      summary: 完整更新产品可变资料
      security: [{sessionCookie: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/ProductIdPath'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateProductRequest'}
      responses:
        '200':
          description: 更新后的产品
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/products/{productId}/modes/{dataMode}:
    put:
      tags: [Catalog, Operations]
      summary: 创建或更新产品模式策略
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
        - $ref: '#/components/parameters/DataModePath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConfigureModeRequest'}
      responses:
        '200':
          description: 更新后的模式策略
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProductMode'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/products/{productId}/publish:
    post:
      tags: [Catalog, Operations]
      summary: 首次发布或重新发布产品
      security: [{sessionCookie: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/ProductIdPath'}]
      responses:
        '200':
          description: PUBLISHED 产品；重新发布保留首次 publishedAt
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/products/{productId}/disable:
    post:
      tags: [Catalog, Operations]
      summary: 停用已发布产品
      security: [{sessionCookie: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/ProductIdPath'}]
      responses:
        '200':
          description: DISABLED 产品
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DAYUSESSID
      description: HttpOnly、SameSite=Lax；生产 HTTPS 下 Secure=true
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 必须与当前匿名或认证 Session 匹配
  parameters:
...
    ProductCodeQuery:
      name: productCode
      in: query
      required: true
      schema: {$ref: '#/components/schemas/ProductCode'}
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    DataModePath:
      name: dataMode
      in: path
      required: true
      schema: {$ref: '#/components/schemas/DataMode'}
    DataModeQuery:
      name: dataMode
      in: query
      required: true
      schema: {$ref: '#/components/schemas/DataMode'}
    FromRequired:
      name: from
      in: query
      required: true
      schema: {$ref: '#/components/schemas/UtcInstant'}
    ToRequired:
      name: to
      in: query
      required: true
      schema: {$ref: '#/components/schemas/UtcInstant'}
...
    CycleTimeOptional:
      name: cycleTime
      in: query
      schema: {$ref: '#/components/schemas/UtcInstant'}
    LeadMinutesOptional:
      name: leadMinutes
      in: query
      schema: {type: integer, minimum: 0}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: pageSize
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    Unauthenticated:
      description: 未登录或认证已失效
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: 无权限、CSRF 失败或路径安全校验失败
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: 产品、资产或事件不存在；公开接口也用它隐藏未发布产品
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    Conflict:
      description: 唯一性、状态机或并发任务冲突
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationFailed:
      description: 参数格式正确但不满足业务校验
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    InternalError:
      description: 未预期系统错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
    ProductCode: {type: string, pattern: '^[A-Z][A-Z0-9_]{1,63}$'}
    UtcInstant: {type: string, format: date-time, pattern: 'Z$'}
    DataMode: {type: string, enum: [REALTIME, FORECAST]}
    AssetType: {type: string, enum: [WEBP, NETCDF]}
    AssetStatus: {type: string, enum: [AVAILABLE, MISSING]}
    ProductStatus: {type: string, enum: [DRAFT, PUBLISHED, DISABLED]}
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: {type: string}
        message: {type: string}
        traceId: {type: string}
        details: {type: object, additionalProperties: true}
    PublicProductMode:
      type: object
      required: [dataMode, staleAfterMinutes]
      properties:
        dataMode: {$ref: '#/components/schemas/DataMode'}
        staleAfterMinutes: {type: integer, minimum: 1}
    PublicProductSummary:
      type: object
      required: [productId, code, nameZh, nameEn, family, producer, colorbarRequired, sortOrder, modes]
      properties:
        productId: {type: integer, format: int64}
        code: {$ref: '#/components/schemas/ProductCode'}
        nameZh: {type: string}
        nameEn: {type: string}
        family: {type: string}
        unit: {type: [string, 'null']}
        producer: {type: string}
        algorithmName: {type: [string, 'null']}
        colorbarRequired: {type: boolean}
        sortOrder: {type: integer}
        modes: {type: array, items: {$ref: '#/components/schemas/PublicProductMode'}}
    PublicProductDetail:
      allOf:
        - $ref: '#/components/schemas/PublicProductSummary'
        - type: object
          required: [descriptionZh, descriptionEn, sourceDescription]
          properties:
            descriptionZh: {type: string}
            descriptionEn: {type: string}
            sourceDescription: {type: string}
            officialSourceUrl: {type: [string, 'null'], format: uri}
            colorbarUrl: {type: [string, 'null'], format: uri-reference}
    AdminProductMode:
      type: object
      required: [dataMode, enabled, staleAfterMinutes]
      properties:
        dataMode: {$ref: '#/components/schemas/DataMode'}
        enabled: {type: boolean}
        staleAfterMinutes: {type: integer, minimum: 1}
    AdminProduct:
      allOf:
        - $ref: '#/components/schemas/PublicProductDetail'
        - type: object
          required: [status, modes, createdAt, updatedAt]
          properties:
            status: {$ref: '#/components/schemas/ProductStatus'}
            modes: {type: array, items: {$ref: '#/components/schemas/AdminProductMode'}}
            publishedAt: {oneOf: [{$ref: '#/components/schemas/UtcInstant'}, {type: 'null'}]}
            createdAt: {$ref: '#/components/schemas/UtcInstant'}
            updatedAt: {$ref: '#/components/schemas/UtcInstant'}
    CreateProductRequest:
      type: object
      required: [code, family, nameZh, nameEn, descriptionZh, descriptionEn, producer, sourceDescription, colorbarRequired, sortOrder]
      properties:
        code: {$ref: '#/components/schemas/ProductCode'}
        family: {type: string, minLength: 1, maxLength: 64}
        nameZh: {type: string, minLength: 2, maxLength: 255}
        nameEn: {type: string, minLength: 2, maxLength: 255}
        unit: {type: [string, 'null'], maxLength: 64}
        descriptionZh: {type: string}
        descriptionEn: {type: string}
        producer: {type: string, minLength: 1, maxLength: 255}
        algorithmName: {type: [string, 'null'], maxLength: 255}
        sourceDescription: {type: string}
        officialSourceUrl: {type: [string, 'null'], format: uri}
        colorbarRequired: {type: boolean}
        colorbarPath: {type: [string, 'null'], maxLength: 512}
        sortOrder: {type: integer}
    UpdateProductRequest:
      type: object
      required: [nameZh, nameEn, descriptionZh, descriptionEn, producer, sourceDescription, colorbarRequired, sortOrder]
      properties:
        nameZh: {type: string, minLength: 2, maxLength: 255}
        nameEn: {type: string, minLength: 2, maxLength: 255}
        unit: {type: [string, 'null'], maxLength: 64}
        descriptionZh: {type: string}
        descriptionEn: {type: string}
        producer: {type: string}
        algorithmName: {type: [string, 'null']}
        sourceDescription: {type: string}
        officialSourceUrl: {type: [string, 'null'], format: uri}
        colorbarRequired: {type: boolean}
        colorbarPath: {type: [string, 'null']}
        sortOrder: {type: integer}
    ConfigureModeRequest:
      type: object
      required: [enabled, staleAfterMinutes]
      properties:
        enabled: {type: boolean}
        staleAfterMinutes: {type: integer, minimum: 10, maximum: 10080}
    PreviewFrame:
      type: object
      required: [webpAssetId, previewUrl, validTime, downloadAvailable]
      properties:
        webpAssetId: {type: integer, format: int64}
        previewUrl: {type: string, format: uri-reference}
        validTime: {$ref: '#/components/schemas/UtcInstant'}
        cycleTime: {oneOf: [{$ref: '#/components/schemas/UtcInstant'}, {type: 'null'}]}
        leadMinutes: {type: [integer, 'null'], minimum: 0}
        downloadAvailable: {type: boolean}
    ForecastCycle:
      type: object
      required: [cycleTime, firstValidTime, lastValidTime, frameCount]
      properties:
        cycleTime: {$ref: '#/components/schemas/UtcInstant'}
        firstValidTime: {$ref: '#/components/schemas/UtcInstant'}
        lastValidTime: {$ref: '#/components/schemas/UtcInstant'}
        frameCount: {type: integer, minimum: 1}
    ScientificAsset:
      type: object
      required: [assetId, fileName, fileSize, products, dataMode, validTime, status]
      properties:
        assetId: {type: integer, format: int64}
        fileName: {type: string}
        fileSize: {type: integer, format: int64, minimum: 0}
        products: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/ProductCode'}}
        dataMode: {$ref: '#/components/schemas/DataMode'}
        cycleTime: {oneOf: [{$ref: '#/components/schemas/UtcInstant'}, {type: 'null'}]}
        validTime: {$ref: '#/components/schemas/UtcInstant'}
        leadMinutes: {type: [integer, 'null'], minimum: 0}
        status: {$ref: '#/components/schemas/AssetStatus'}
    ScientificAssetPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/ScientificAsset'}}
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, format: int64, minimum: 0}
...